feat(json-serializer): do not treat `all` as a special segment group
We treat it like any other group to populate a section list.
Instead, we create a `nrn_all` reserved section list that contains
all sections in it. If a user creates a `nrn_all` segment group in their
NeuroML representation, we throw an exception and ask them to use a
different name.

Fixes #99