feat(neuron/netpyne): use `%r` instead of `%e` or `%s`
when writing recorded data to file.

`%r` or the `repr` function gives the "official" string representation
of an object, so of a float too, while the `%s` or `str` method returns
the "nicely printable string representation of an object".

Also, irrespective of which one we use, we must use the same for both
NEURON and NetPyNE.